Fixes #269 - break up gedrenderer
* Move the management of sub-renderers to a new base class
